Update colophon.xhtml
Changed the date of publication from 1898 to 1891. 

In the colophon (https://standardebooks.org/ebooks/selma-lagerlof/the-story-of-gosta-berling/pauline-bancroft-flach/text/colophon), it said: 
The Story of Gösta Berling
was published in 1898 by
Selma Lagerlöf.
It was translated from Swedish in 1898 by Pauline Bancroft Flach.

While on Wikipedia and on the book's page itself on Standard Ebooks (https://standardebooks.org/ebooks/selma-lagerlof/the-story-of-gosta-berling/pauline-bancroft-flach), it said 'The novel was published in Swedish in 1891'.